remove a bunch of uncessary casts and other warnings from eclipse
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@3687 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/io/object/UnsignedByte.java
@@ -77,12 +77,12 @@ public final class UnsignedByte extends UnsignedNumber {
 
 	@Override
 	public double doubleValue() {
-		return (double) value;
+		return value;
 	}
 
 	@Override
 	public float floatValue() {
-		return (float) value;
+		return value;
 	}
 
 	@Override
@@ -123,7 +123,7 @@ public final class UnsignedByte extends UnsignedNumber {
 	
 	@Override
 	public int hashCode() { 
-		return (int)value;
+		return value;
 	}
 
 	@Override
